issue-5895: updated silk to bef85a568e4f5dd3e71e9557ec190972514ecb05

+++ contrib/libs/silk/src/fibers/cpu.cpp
@@ -36,7 +36,7 @@ static int readSysfsUint32(const char * path, uint32_t * out) noexcept
     buf[n] = '\0';
 
     char * end;
-    uint32_t val = ::strtoul(buf, &end, 10);
+    uint32_t val = static_cast<uint32_t>(::strtoul(buf, &end, 10));
     if (end == buf)
     {
         return EINVAL;
